Enhanced file structure, refactored slightly
