+++ app/src/main/java/com/example/roomer/data/local/RoomerStoreInterface.kt
@@ -1,4 +1,4 @@
-package com.example.roomer.data.repository
+package com.example.roomer.data.local
 
 import com.example.roomer.domain.model.entities.Room
 import com.example.roomer.domain.model.entities.User
@@ -12,7 +12,7 @@ interface RoomerStoreInterface {
     suspend fun deleteFavourite(room: Room)
     fun getCurrentUser(): User
     suspend fun deleteCurrentUser()
-    fun getAllUsers(): List<User>
+    fun getAllUsers(): Flow<List<User>>
     suspend fun deleteUser(user: User)
     fun getUserById(userId: Int): User
     suspend fun addUser(user: User)
