+++ app/src/main/java/com/example/roomer/di/AppModule.kt
@@ -1,7 +1,11 @@
 package com.example.roomer.di
 
+import com.example.roomer.data.local.RoomerStoreInterface
 import com.example.roomer.data.remote.RoomerApi
-import com.example.roomer.data.repository.*
+import com.example.roomer.data.repository.auth_repository.AuthRepository
+import com.example.roomer.data.repository.auth_repository.AuthRepositoryInterface
+import com.example.roomer.data.repository.roomer_repository.RoomerRepository
+import com.example.roomer.data.repository.roomer_repository.RoomerRepositoryInterface
 import com.example.roomer.utils.Constants.BASE_URL
 import dagger.Module
 import dagger.Provides
