+++ app/src/main/java/com/example/roomer/domain/usecase/login_sign_up/SplashScreenUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.roomer.domain.usecase.login_sign_up
 
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.roomer_repository.RoomerRepositoryInterface
 import com.example.roomer.utils.Constants
 import com.example.roomer.utils.Resource
 import java.io.IOException
