+++ app/src/main/java/com/example/roomer/domain/usecase/shared_screens/ChatUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.roomer.domain.usecase.shared_screens
 
-import com.example.roomer.data.repository.RoomerRepository
+import com.example.roomer.data.repository.roomer_repository.RoomerRepository
 
 class ChatUseCase(
     private val roomerRepository: RoomerRepository,
