+++ app/src/main/java/com/example/roomer/presentation/screens/entrance/signup/interests_screen/InterestsScreenViewModel.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.roomer.data.repository.AuthRepositoryInterface
+import com.example.roomer.data.repository.auth_repository.AuthRepositoryInterface
 import com.example.roomer.domain.model.login_sign_up.interests.InterestModel
 import com.example.roomer.domain.usecase.login_sign_up.SignUpUseCase
 import com.example.roomer.utils.Resource
