+++ app/src/main/java/com/example/roomer/presentation/screens/entrance/splash_screen/SplashScreenViewModel.kt
@@ -3,7 +3,7 @@ package com.example.roomer.presentation.screens.entrance.splash_screen
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.roomer_repository.RoomerRepositoryInterface
 import com.example.roomer.domain.usecase.login_sign_up.SplashScreenUseCase
 import com.example.roomer.presentation.screens.UsualScreenState
 import com.example.roomer.utils.Resource
