+++ app/src/main/java/com/example/roomer/presentation/screens/navbar_screens/MessengerViewModel.kt
@@ -5,8 +5,8 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.roomer.local.converters.convertTimeDateFromBackend
-import com.example.roomer.data.repository.RoomerRepository
+import com.example.roomer.utils.converters.convertTimeDateFromBackend
+import com.example.roomer.data.repository.roomer_repository.RoomerRepository
 import com.example.roomer.domain.model.entities.Message
 import com.example.roomer.domain.usecase.navbar_screens.MessengerUseCase
 import com.example.roomer.presentation.screens.entrance.login.LoginScreenViewModel
