+++ app/src/main/java/com/example/roomer/presentation/screens/navbar_screens/home_screen/HomeScreenViewModel.kt
@@ -2,7 +2,7 @@ package com.example.roomer.presentation.screens.navbar_screens.home_screen
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.roomer.data.repository.RoomerRepositoryInterface
+import com.example.roomer.data.repository.roomer_repository.RoomerRepositoryInterface
 import com.example.roomer.domain.model.entities.Room
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
