+++ app/src/main/java/com/example/roomer/presentation/screens/shared_screens/ChatScreenViewModel.kt
@@ -6,10 +6,10 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.roomer.local.converters.createJson
-import com.example.roomer.local.converters.getFromJson
+import com.example.roomer.utils.converters.createJson
+import com.example.roomer.utils.converters.getFromJson
 import com.example.roomer.data.remote.ChatClientWebSocket
-import com.example.roomer.data.repository.RoomerRepository
+import com.example.roomer.data.repository.roomer_repository.RoomerRepository
 import com.example.roomer.domain.model.entities.Message
 import com.example.roomer.domain.model.entities.User
 import com.example.roomer.presentation.screens.entrance.login.LoginScreenViewModel
