+++ app/src/main/java/com/example/roomer/room/RoomerDatabase.kt
@@ -1,11 +1,11 @@
-package com.example.roomer.local
+package com.example.roomer.room
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import com.example.roomer.domain.model.entities.User
-import com.example.roomer.local.dao.UserDao
-import com.example.roomer.local.dao.FavouriteDao
-import com.example.roomer.local.entities.LocalRoom
+import com.example.roomer.room.dao.UserDao
+import com.example.roomer.room.dao.FavouriteDao
+import com.example.roomer.room.entities.LocalRoom
 
 @Database(entities = [LocalRoom::class, User::class], version = 1)
 abstract class RoomerDatabase: RoomDatabase() {
