+++ app/src/main/java/com/example/roomer/utils/converters/date_time_converter.kt
@@ -1,4 +1,4 @@
-package com.example.roomer.local.converters
+package com.example.roomer.utils.converters
 
 fun convertTimeDateFromBackend(rawData: String): String {
     return rawData.substring(5, rawData.indexOf('T'))
